improve update progress status messages
before this change, the update progress counter displayed the number of feeds
added to the update queue. this is not what most users would expect to see, as
it would be more obvious to display the number of completed updates.
also, since in the vast majority of cases multiple feeds are updated
simultaneously, it's awkward to display just a single URL, so it's removed