continue work on HTML formatting edge cases
in this refactoring i handed over the responsibility of drawing indentation for
render lines to the pager_list_menu_writer function, because to make formatting
work with constant consideration of these indentations is a bloody hell...

also there was a case when formatting started before indentation is created,
which could make white space appear underlined at the beginning of a line.
there were certainly ways to handle this case, but this would have made the
already very bloated code smell even more...

therefore at this point it is impossible to give any formatting to the
indentation at the beginning of lines, so be it.