add pager-centering setting (#89)
To have the pager aligned to the left even when pager-width is > 0.
In large terminals I find it useful to leave the pager-width > 0 and to be able to keep the feed text aligned to the left, where also aligns all the newsraft menus.
It defaults to `true`, so that the behaviour prior to this commit is unchanged.

Reviewed-on: https://codeberg.org/newsraft/newsraft/pulls/89
Co-authored-by: _fe <ykkl@posteo.net>
Co-committed-by: _fe <ykkl@posteo.net>